+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 * EPMRPP-25178 - limited minimum height (415px) for application in desktop view.
 * EPMRPP-25373 - Open link in a new browser tab by Ctrl+click combination
 * EPMRPP-25182 - fixed backlinks after bug posting.
+* EPMRPP-25348 - fixed email validation for invite user modal